Duck functions update
spawn() function was updated to always start from the grass, so it is unable to spawn in the sky.
Added a flyAway() function, which makes the duck exit the screen to the right or left depending on its position.
Added foreground for the duck to hide behind.
Fixed distances according to screen size.